Use spinalcordtoolbox/report-ui
Updates the QC report feature to use
https://github.com/spinalcordtoolbox/report-ui - a ground-up rewrite
of the frontend. For all intents and purposes, nothing should have
changed from a user perspective; the new UI was made to be fully
backward-compatible with the old one.

A forthcoming commit will document the procedure for updating to
new versions of report-ui