Add `-timeit` argument to "common args" argparse group to enable profiling for CLI scripts (#4825)
## Description

We've been wanting to be able evaluate the run-time of our programs for
a while. This PR aims to define a manner for this to be done in a
standardized way, while making it available to any and all CLI
interfaces which wish to utilize it through `argparse` actions.

This PR focuses on time-based profiling, measuring and reporting the
total runtime _only_ of any CLI program which uses the `-timeit` flag.
It will also report the runtime if the program is interrupted in any way
(see "Known Issues" below for caveats).

If there is interest, we can move to a more thorough (though
heavy-weight) profiling scheme via the
[`profile`](https://docs.python.org/3/library/profile.html#module-cProfile)
tool, but I want to test the impact on total runtime before doing so; as
the fact the profiler itself adds to runtime when run "in-line", we need
to assess that impact before deciding which method to use. In the
meanwhile, consider using the package directly (which is known to have
minimal performance impact)

## Known Issues

Due to an upstream issue with how `atexit` is implemented (using "weak"
references to the functions its registered to run at-exit), if the
program crashes due to an error, the runtime is printed in the middle of
the stack-trace, rather than the end.

If the program is terminated externally (i.e. running out of time on a
SLURM-mediated job) any `atexit` functions will not be run, including
the profiling call used in this PR. This is true of literally any code,
however, but is something to keep in mind if you're stress-testing
things...

## Linked issues

Resolves #4801

---------

Co-authored-by: Kalum Ost <kalum.ost@polymtl.ca>
Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>