Revert contrast-agnostic `spinalcord` model back to v2.5
Previously, in #4783, we merged the `v3.1` nnUNet model. But, it turns out that versions `v3.0` and `v3.1` were identical, and also were named inconsistently (making it hard for Naga to test).

In the short term, Naga has deleted the v3.0 and v3.1 releases and is remaking them as one singular v3.0 model. However, this is causing tests to fail on SCT's master branch. So, this PR reverts back to the last working model temporarily until the v3.0 model can be re-released.

I'm merging this directly to master because this is a fully temporary change (there will be a proper follow-up PR), and I want our CI runs to pass while we wait for a new v3.0 release as to not disrupt ongoing PRs.

Partially addresses https://github.com/sct-pipeline/contrast-agnostic-softseg-spinalcord/issues/142#issuecomment-2770983399.