`centerline/core.py`: Move `-r` step to the end of the function
The bug in this PR stems from some confusing behavior between how we handle `im_centerline` and `arr_ctl`.
I want to remove as much ambiguity and mental load from this function as possible. So, by simplifying the
`-r` handling code, it will make the latter part of the function a lot easier to understand.