`types.py`: Improve documentation of `Coordinate.permute` method
This method is useful for properly reorienting the coordinates. However, it's a little messy right now, so
we might as well clean it up first by removing an unused parameter and filling in the docstring.

Note: We're forced to fix a linting issue in this commit to placate the linter. ;_;