`centerline/core.py`: Make reorientation fix opt-in (for now)
This will allow us to better test the impact on the callers of this function, with and without the fix.

For now, I've only set `reorient_array=True` for `sct_get_centerline` (i.e. the output array). But, ideally
we would be able to set it to `True` (or remove the option entirely), and set reorientation on by default.