`types.py`: Fix off-by-one error
This makes sense, right? Say we have an image with dim = 20. If the original coordinate is 0, the inversion
of that should be 19, since dim=20 -> [0, 19].

The only reason I spotted this off by one error is because in the RPI -> SAL test, all three axes get
inverted, and the coordinates were off by one for all three axes!!

I guess that's one reason to re-use code, as it allows me to spot this pre-existing bug!