Fix slice indexing for `-angle-corr-centerline` (#4850)
Fixes #4807.

If the input mask supplied in `-i` and the custom centerline supplied in
`-angle-corr-centerline` did not start at the same lowest SI slice, then
the wrong slice index was used for angle correction in the old code.

With the new code, everything is indexed directly with SI slice
numbers, rather than trying to do relative indexing. A quick manual test
using the data from #4807 shows that this fixes the problem: for the
slices which appear in both CSVs, they have the same `angle_AP` and
`angle_RL`.

If the `-angle-corr-centerline` is too short, there's now an error
message to explain the situation.

---------

Co-authored-by: Kalum Ost <43527660+SomeoneInParticular@users.noreply.github.com>