Refactor QC report UI (#4827)
## Description

To allow for easier updates and bug fixes on QC reports, the
javascript/HTML has been extracted to
[spinalcordtoolbox/report-ui](https://github.com/spinalcordtoolbox/report-ui),
where it can take advantage of a build system meant for JavaScript, as
well as other modern tooling, without adding noise to this repo.
Instead, version releases from there will be copied to this repo when
appropriate (see 85237653a876a71106c2d2ea05c62367d47300f6 for an example
update, with instructions).

The current implementation aims for full backward compatibility with the
current report system. That said, since generated reports are
self-contained HTML directories, some minor changes were convenient to
make, namely:

- the JSON generated to aid rendering the report uses camelCase for
better compatibility with JS naming standards, and some fields are
renamed for clarity
- the directory structure of the generated reports has changed
- the YAML created by saving fails/artifacts has been slightly
streamlined

The new UI comes with many performance and layout improvements while
still maintaining the structure & feel of the legacy UI. Users with
muscle memory for the reports should find it effectively the same.

An upshot of this rewrite is that new features, refactors, and fixes
should be much easier: the modularity and pure-function orientation of
React makes this so!

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>