Pass `-step1-only` to `totalspineseg` by means of a `**kwargs` pattern
Because we use a common function signature between all the inference functions:

   inference(path_img, tmpdir, predictor, device)

It is a little awkward to add an inference argument that is only relevant for one framework (totalspineseg)
but not the other frameworks (nnunet, monai).

To accomplish this, we use a dictionary of "extra" keyword arguments to smuggle in kwargs that may be
present for some tasks/frameworks/subparsers, but not for others.

It's possible that this is a smell that the "common function signature" has outgrown its usefulness, and
that we need to refactor how arguments are parsed for `sct_deepseg`? But this works for a first draft,
and I like that it clearly delineates between common arguments and unique arguments.