Add `-step1-only` argument to the `totalspineseg` task's subparser (#4860)
## Description

This PR adds a new argument `-step1-only` to `totalspineseg`'s
subparser, which gets passed to the `step1_only` parameter of the
`inference()` function for `totalspineseg`. This allows users to skip
step2, which may be costly/time-consuming.

Additionally, this PR adds a test, with the resulting QC report shown
below:

[Screencast from 2025-04-16
11-14-57.webm](https://github.com/user-attachments/assets/bd4a6f4e-f8a9-4273-96e8-183c3284179a)

## Linked issues

Fixes #4794.

---------

Co-authored-by: Kalum Ost <43527660+SomeoneInParticular@users.noreply.github.com>