`test_sct_deepseg.py`: Parametrize `expected_dice` (to support `0.91`)
Since the graymatter model has a lower expected Dice score, and since we've already made exceptions for
a different model in the past, it seems like requiring the same Dice for all models is prohibitive.

So, this commit adds `expected_dice` to the test parametrization, then allows each model to specify its
own expected Dice score.