`qc2.py`: Factor out canvas size constants to enforce a fixed width
I think the requirement that the canvas width was fixed was not very well understood, and so various
QC reports were playing with the canvas size, not considering the target width of 1500px, etc.

By defining these as constants, we communicate that there is an explicit assumption being made here,
and either QC reports will adhere to the assumption, or they'll challenge the assumption (which will
be informative for how we design the output images).