`qc2.py`: Fold integer scaling into pre-existing float scaling
Previously, when adding the `scale` feature, we hadn't realized that there was already scaling going on!
   > (setting the array size smaller than the canvas size will cause matplotlib to upscale)

And, in fact, the existing `matplotlib` scaling is nicer, since it doesn't require an integer scale factor.

Given that we now know of this existing scaling, there is no need to use an entirely separate mechanism
to scale the mosaic up. Instead, we can keep each mosaic tile at its original size, and scale the "max width"
by the same amount instead, causing `matplotlib` to eventually scale the array to fit the canvas instead.

Note: Since we are keeping the smaller "original size" mosaic array, we no longer need to scale the radius
      when adding orientation labels, either.