`qc2.py`: Fix bug with how `max()` is used to determine scaling
Take this sample anisotropic image: [0.6, 0.8]
   - The ratios relative to the human baseline of 0.6 will be [1.0, 0.75]
   - The scaling factors, then, will be [1.0, 1.33]
   - So, the expected scaling would be 1.33.

However, because `max()` was performed _before_ the division by 1, then the minimum would be used (1.0).
So, we just move the `max()` call to happen after the division.
(We could have just switched to `min()` prior to dividing, but I wanted to preserve the intent, and `max()`
communicates that better.)