Add fixes for `PyQt5` and `onnxruntime` to amend our previous switch to Python 3.10 (#4906)
## Description

In PR #4869, we switched our conda environment from Python 3.9 to Python
3.10. While the tests passed on the PR, they failed for the
`master`-only tests: #4905.

This PR tries to address 2 separate bugs from that issue:

- `PyQt5==5.15.x` has a finicky sister package `PyQt5-Qt5`, where only a
single version is cross-platform.
- `onnxruntime` for Python 3.10 requires a newer version of
`msvc-runtime` than what we had been pinning.

To address the former, we pin to the single working version, and
subsequently update our hacky check of PyQt5 versions.

To address the latter, we update the version of `msvc-runtime`.

## Linked issues

Fixes #4905.