`check-url.sh`: Add a blanket exception for 403 - Forbidden responses
This should save some developer headache when a future site chooses to ban automated scrapers.

I don't think this will cause any extra broken links to slip through? If we're getting 403, then presumably
it will be impossible to gain any sort of useful signal from the site regardless of whether the url is
valid or invalid.