Add a blanket exclusion for `403 - Forbidden` responses to our broken link checker (#4908)
## Description

Because of the rise of AI/LLM web scrapers, more and more sites are
forbidding access to automated checks. It's quite tedious to manually
blacklist each site every time this happens. Sure, there's a small
chance that we could manually find an alternate URL for a 403'ing URL.
But, for most cases, if a site is giving us 403 Forbidden, there's not
much we'll be able to do.

In theory, we could set up a system where we continue to blacklist
domains, but manually check the URLs once they get blacklisted (so we at
least know which domains to manually test). But, I fear that this is a
lot of effort for a relatively minor nuisance. So, I'm content to just
filter out 403 domains for now?

## Linked issues

Fixes #4897.