Remove default value from Apptainer bundling script to fix CI error (#4911)
## Description

Typically, automated (non-manual) runs of the `create-release.yml`
workflow have used an empty (null) value for `milestone_title`. This has
worked okay, since all of the steps are compatible with having an empty
string in that place.

However, in #4857, we added a new script that bundles the Apptainer
files as well. And, this bundling script will fallback to a default
value of 7.0. This works in the context of the script, but in the GitHub
Actions workflow context, the default is still `""`, meaning the 7.0
file can't be found. (Failing run:
https://github.com/spinalcordtoolbox/spinalcordtoolbox/actions/runs/15345371948/job/43180335856#step:6:14)

This PR removes the default value to allow it to run with `""` as per
the norm.

> [!NOTE]
> Ideally, instead of allowing an empty version (""), we would set a
default version in the outer GitHub Actions workflow context. However,
GitHub makes it hard on us -- we can't just use the `default:` field of
the workflow_dispatch trigger:
https://github.com/orgs/community/discussions/26322
>
> So, instead, we begrudgingly allow empty values for `$SCT_VERSION`.

## Linked issues

Amends #4857.
Fixes #4593.