`algorithms.py`: Use variables instead of strings for im/seg filenames
Previously, we used the filenames "src.nii" and "dest.nii" for "rot_method=pca". However, this is not
safe, because these filenames could refer to either the `im` or `seg`, but only `seg` is valid.

To alleviate this issue, we use variables for both `im` and `seg`, to ensure that only `seg` images will be
used for `rot_method=pca`.

The problem with this fix is that, because the same filename is used, the user could theoretically provide
`type=im` and end up with suboptimal results. But, this was a problem with the old code, too!

At the very least, this is a minimal change that improves clarity in the code and fixes the `imseg` case,
without making a mess of the other cases. So, it feels safe.