`registration/core.py`: Validate `type` used for each `rot_method`
This fix sort of renders the previous commit moot. But, due to the complex structure of our registration
functions, I wasn't 100% sure that all the code paths were covered with either fix, hence including both.

- sct_register_to_template -> register_wrapper -> register -> register_step_slicewise -> register_slicewise
- sct_register_multimodal  -> register_wrapper -> register -> register_step_slicewise -> register_slicewise

Maybe this should be in `register_wrapper` instead? Just because putting in `register()` could cause it
to fail at a later step (wasting processing). But, I like that by including it here, we get to display the
step parameters back to the user so that they have the full context. This is because `pca` is the default
rot_method, so the error may be caused by a default parameter (i.e. one not specified by the user). And thus,
if we validate the inputs immediately, they may be confused as to why they're getting an error for a
parameter they didn't even specify.