`sct_apply_transfo`: Fix how cropping dimensions are determined
If we plan to set everything outside the warping field to 0, then we should be using the warping field's shape
with the warping field's header, not the src image's dimensions!

This fixes the "weird bounding box" behavior seen in https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4876#issuecomment-2960122864

Also, the [:, :, :] syntax doesn't even do anything! It will preserve the 4th/5th dimensions of the warping
field. So, instead of that, since we're trying to make a mask anyway, why not just create a new array of ones
 from the warping field's dimensions. This also lets us nix the `if dim == '2'` code.