Add safeguards to handle when `type=imseg` is used with `rot_method=pca` (#4924)
If `type=imseg` is supplied with `rot_method=pca`, the wrong image
(`im`) will be used instead of the required image (`seg`). This is due
to some ambiguous filenaming conventions, which we fix in commit
2c21ae1f4fcbb15c6685fc0ef645df5a85903ca6.

However, that fix still allows the possibility for the user to
erroneously specify `type=im` instead. So, we go even further and add an
explicit check for each step to enforce `type=seg` for `rot_method=pca`.

Perhaps only the latter change is needed? But, I liked the idea of
having redundancy, in the case that there's some other entry point to
`register_slicewise()` besides the core `register()` function.

## Alternative solutions

Whether or not these are the correct fixes is totally up for discussion.
But, I wanted to make minimal changes that didn't require a large
refactor, e.g.:

- Explicitly passing `type` to the `register_slicewise` function (many
layers deep!)
- Parsing the input filenames for `_seg` to determine filetype
(brittle!)
- Updating the `register_slicewise()` arguments to have explicit values
for `fname_src_im`, `fname_src_seg`, etc.