`straightening.py`: Clarify that `bound_straight` == `[start, end]`
Right now, the variable `bound_straight` is used in two places:

1. As the "safe zone" for the straight space.
2. As the start/end points for the straight space (when we don't have a reference image)

We have determined via experimentation that these values are not actually a valid safe zone. However, they
_are_ still useful for creating the straight space when we don't have a reference image.

Because we plan to rewrite the "safe zone" logic, but leave the "start/end points" logic untouched, we can
rename some variables (and move the logging output) so that it's clearer that these values are only relevant
for the "start/end points" logic.