`straightening.py`: Fix off-by-one indexing error
Since the `z:` slice syntax is inclusive of `z`, we need to increment by 1 to not include that slice.

Additionally, I updated `0:z` syntax to be `:z` instead. This has no functional difference, but it matches
the indexing syntax of `z:`, and I mention both in the comment, so I figure it'll make things easier to grok.