`install_sct`: Only validate the default directory once
If the default directory is invalid, a bunch of warnings will be printed.

If the user then enters an alternate install directory that is _also_ invalid, then those same "default"
warnings will be printed all over again.

This is quite noisy, and I think unnecessary? So, we should only run the validation on the first iteration
of the loop. Then, if the user enters something invalid after the fact, we jump straight into prompting them
again, skipping the entire "default dir validation" part.