`install_sct.bat`: Move "folder existence" warning
Since the user has no opportunity to change the path if it already exists, then the warning should be moved
alongside the other existence check.

Perhaps we should prompt the user if the directory already exists? But I figure that because the installer
script doesn't have a "non-interactive" mode, the directory will always be explicitly chosen by the user,
meaning we're less likely to destroy sensitive directories. (Besides, only specific directories are removed.)