Mask output of `sct_apply_transfo` when applying `-initwarp` during multimodal registration (#4927)
## Context

Registration (e.g. `sct_register_to_template`) will produce warping
fields that map the voxels in one coordinate space (e.g. the PAM50
template space) onto the space of another image.

If a warping field is originally determined using an image with one FOV
(e.g. a T2 image with dimensions (40, 515, 512)), but then the warping
field is applied to an image with a different FOV (e.g. a dMRI image
with dimensions (256, 256, 13)), then the ANTs tool
`antsApplyTransforms` may hallucinate phantom copies of the image in
areas where the initial warping field was not valid.

When would a warping field be applied to an image with a different FOV?
Well, in fact, we do it all the time when we specify the `-initwarp` and
`-initwarpinv` arguments [as a "pre-registration"
step](https://spinalcordtoolbox.com/stable/user_section/tutorials/diffusion-weighted-mri/template-registration-for-dmri.html#registering-the-template-to-the-dti-space).

(Side note: More often than not, this will affect `-initwarp`
specifically, because the template -> anat will have different anat
FOVs. `-initwarpinv` is spared, because the anat -> template uses the
same destination space (the template) in both cases. However, in the
case of e.g. multimodal non-template registration, `-ref template`, etc.
this assumption may not hold.)

## Description

To mitigate this issue, because we have the warping field, it's possible
to mask out any output voxels beyond the warping field FOV (considering
them "invalid"), which removes this phantom copy:

And as it turns out, `sct_apply_transfo` has an option called `-crop`
which was already designed to perform this exact masking step, added all
the way back in 2015. However, `-crop` is broken in several ways:

- It uses the wrong mask when cropping the image (data dimensions,
rather than warping field dimensions)
- It uses a rectangular bounding box derived from the min/max coords of
the warping field. This acts as a sort of [convex
hull](https://scikit-image.org/docs/0.24.x/auto_examples/edges/plot_convex_hull.html),
covering an area that's a little larger than the warping field itself,
which may be too large and not mask out enough voxels.

To address these two issues, this PR adds the following fixes:

- Use the correct dimensions when creating the mask to use for `-crop`.
- Adds a new option to the `-crop` flag (`-crop 3`) to mask out voxels
instead of just cropping with a rectangular bbox.

While `-crop 3` will solve the issue for user calls to
`sct_apply_transfo`, it won't actually fix the problem in any scripts
that internally call `sct_apply_transfo`. Because this issue is commonly
seen inside `sct_register_multimodal` with `-initwarp`, I decided to
turn on this flag when `-initwarp` is used (e.g. `step=1`).

I don't think changing this default should have any effect unless these
phantom copies were present, and even then, the effect should only ever
be positive. This should hopefully be confirmed by the fact that the
`batch_processing` results should not change despite the use of
`-initwarp`.

## Alternative approaches

In theory, we could crop the destination image to the same extents as
the warping field instead (_before_ we apply `antsApplyTransform`,
rather than after). Then, we would update the dimensions of the cropped
output image to match the full FOV of the destination image. This would
be more thorough and likely prevent phantom copies from appearing at
all, which might be relevant in situations where the phantom copy
[overlaps](https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4876#issuecomment-2959955246)
the true copy.

Note: The destination image would have to be truly cropped (i.e.
changing the dimensions) and not just masking the voxels to `0`, since
`antsApplyTransform` only cares about the destination FOV and not the
contents of the destination image.

However, this is a more complicated approach that requires some care,
because cropping the image changes the origin of the array (and thus the
affine). I couldn't quite wrap my head around how to do this properly.
(Something like generating the warping field mask, passing it to
`sct_crop_image -mask`, running `antsApplyTransforms` on the cropped
image, then... well, I'm not sure how to fix the dimensions of the
output image. I thought about combining it with a `zeros_like` image,
but `sct_maths -add` imposes a "same dimensions" restriction.)

Because of this, I stuck with the simpler approach of masking the output
instead, which seems to work [well
enough](https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4876#issuecomment-2966803339)
on the "overlapping" phantom copy cases.