`sct_register_to_template`: Turn off "safe_zone" by default
Now, there will be no application of the safe zone unless the user specifies it as an advanced parameter
during `sct_striaghten_spinalcord`.

This means that all of my "safe zone" changes are moot, but that's OK, because I think it still provides
insight into the underlying code and how much of it is (or isn't) relevant.