`straightening.py`: Don't stop at `// 2` slice when deduplicating
By stopping at `len()//2`, these loops would not fully de-duplicate all the slices.

I'm not sure why these `//2` stopping points were put in place, given that the loops will break on their own
if they encounter a pair of slices without duplication. It's very strange, almost like it was designed to
cause this exact bug.

I also added some comments to explain the purpose of these loops, since the effect of setting a lookup table
index to `0` was not clear when reading through the code.