LazyLoad `nibabel` across all of SCT's scripts
This was one of the last big "slow" imports that we deemed acceptable for now. I wanted to try this out to
see the impact, and it seems to be around ~0.6s of time save?

The other big libraries (numpy, scipy, etc.) tend to use their own method of quick-loading, meaning they only
take 0.1-0.2s at most to load, so we really hit diminishing returns there.

It's possible that the time save here isn't worth the added complexity for a (relatively) commonly used
library. So, I can revert this change if desired.