`test_cli_sct_detect_compression`: Simplify indexing to remove warning
Before, we extracted a dataframe (instead of a series), meaning we had to use `.values` then convert a len-1
dataframe to a float. But, there's no need for that! We can just set the index column, then index the
dataframe in a much simpler way, preventing the call to `float()` in the first place.