Address new `DeprecationWarnings` thrown during `pytest` test suite (#4940)
## Description

This PR checks the recent test runs in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/actions/runs/15822478546,
and adds fixes for all of the recent DeprecationWarnings that have been
thrown since the last time we did this maintenance. (I also confirmed
that there are no OS-specific warnings).

Note: This doesn't actually fix <!-- --> #4407, since the numerical
issues are left unaddressed for now, but it is related (since it adds a
new entry to the ignore list).

## Linked issues

Related to #4407.