Validate `$SCT_DIR` for path-length issues (even for the default directory + non-interactive mode) (#4935)
## Description

The SCT installation will fail in an unintuitive way on Ubuntu when the
path length of `$SCT_DIR` exceeds 107:

![image](https://github.com/user-attachments/assets/aa06fad9-4c9b-42c7-be8b-c8854f29ee65)

This PR adds a check for this path length, and provides a more intuitive
error instead:

![image](https://github.com/user-attachments/assets/2fd33437-e5dd-4649-9fc1-7e34213c76cb)

However, while authoring this PR, I encountered a fatal flaw: The path
checks will only be applied for user-entered directories. If the default
directory has issues and the user accepts the default directory, or if
the script is run in non-interactive mode, then the path will not be
checked and the user will be given no chance to fix the issue,
potentially resulting in destructive behavior on unsafe paths.

So, to address this loophole, I reworked the conditional structure to
check the default directory. In addition, this check will escalate from
warning to error in case the user is in non-interactive mode and cannot
fix the issue:

![image](https://github.com/user-attachments/assets/a5eb05c6-372c-417d-8794-24c1d8fe7d2e)

Post-review, these features were added to the Windows installer as well.

## Linked issues

Fixes #4813.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>