Make sure that all useful parameters in `ParamMoco` are documented in the `-h` (#4979)
## Description

The `ParamMoco` class has many arguments, but not all of them are
exposed to the user. The `-param` argument can actually edit all
parameters in the class, but unless those parameters are documented, the
user won't know what they can edit.

So, this PR updates `sct_dmri_moco -param`, `sct_fmri_moco -param`, and
`ParamMoco` to all be consistent. Any parameters that should be editable
are now documented in the argparse `-help` (such as `iter`), and any
parameters that the user _shouldn't_ edit now have inline comments
explaining why. Also, all useful parameters now have entries in the
`ParamMoco` docstring, too.

> [!NOTE]
> Because `ParamMoco` has an `update()` method that will automatically
parse user input and assign it to the class attributes, most of the
changes in this PR are documentation only. There are some functional
changes to the `update()` method, but these are pretty harmless and are
just meant to catch typos from the user.

## Linked issues

Related to #4978.

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>