Implement anisotropic dilation and erosion (#4966)
This allows using the `sct_maths -dilate` and `-erode` operations with a
single number of voxels (as before), or a list of 2 or 3 numbers
separated by `x`, like `1x2x3`, which can be useful for anisotropic
images.

While `skimage.morphology.footprint_rectangle` works for 2D and 3D box
shapes, the `disk` and `ball` functions only take a single radius. The
`ellipse` function could have worked for 2D, but not for 3D, and it
confusingly turns (width, height) arguments into a (height, width)
resulting array. I decided that it was simpler to just generate the
correct numpy array directly.

Fixes #4936.