`qc2.py`: Ditch heuristic-based `scale` method for adding seg outlines
In practice, because of the `base_scaling` values, this would always be on for `rootlets` and always be off
for `totalspineseg`. And in fact, since I increased `base_scaling`, this condition now gets tripped
erroneously for `totalspineseg`.

We already have an `outline` parameter that was never used (`True` by default), meaning that if labels were
added, outlines would _always_ be added too. So, I think that makes it safe to group both functions under the
same condition, repurpose `outline` for this functionality, and set `outline=False` for totalspineseg to
do away with the heuristic.