Tweak `totalspineseg` axial mosaic QC report to make it easier to read (#4993)
When testing the "full `pytest` QC report" check, I noticed that the
`totalspineseg` axial mosaic QC report looked very small and hard to
read.

This PR adds several fixes to make the size a bit more manageable. It
should be easy to double-check the QC report by checking the artifact
uploaded to `tests.yml` for Ubuntu 24.04.

Note that this is still just a stopgap for when we can implement a
better QC report for totalspineseg, as the axial mosaic isn't all that
useful to begin with.