Update and reorganize Docker installation instructions (#5001)
The point of this pull request is to:

- Make Docker installation instruction more visible, by making a
top-level page for them (alongside MacOS, Linux, Windows, and
Apptainer).
- Provide a Dockerfile for those who prefer to use that, rather than
doing an interactive installation and saving the resulting Docker
container.

It turns out that we already _had_ an example Dockerfile, but it wasn't
linked from anywhere in the documentation. Also, the accompanying README
listed a mismatched version of Ubuntu.

Along the way, I fixed a few other typos and issues I found in the files
I was already touching.

Fixes #4922, fixes #4923.

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>