`sct_deepseg`: Rewrite `-o` to remove mention of multiclass suffixes
By specifying `multiclass`, we would have to go through all of our models and denote them as single/multiclass
just to selectively apply the message to some models but not others.

By rewriting the text to be agnostic to the number of classes, we can use a single generic message instead.