`sct_deepseg`: Hide `-thr` argument for models with `None` default
If a model has a default threshold value of `None`, then it means that either the outputs are already
binarized, or the outputs are multiclass. In both cases, the user should not change the value of `-thr`.

By limiting the argument to only models with `thr`, we can display the default value, too!

Note: This required manually adding the default values to ivadomed. Normally, these values live in the model
folder. This runs the risk of the documentation value and the on-disk value being out of sync. But, because
all the default thr values are either N/A or 0.5, and because the ivadomed project is dissolved, it is 99%
likely that we never train another ivadomed model again. So, I think the duplication is okay here.