Clean up argparse help to only display arguments and descriptions relevant to specific tasks (#5008)
## Description

This PR aims to improve the `sct_deepseg -h` (argparse) by limiting what
information gets displayed for individual task help (e.g. `sct_deepseg
totalspineseg -h`).

The goal is to try and remove arguments that are irrelevant for certain
tasks, and also limit warnings and notes to only the tasks that they are
relevant for.

Here is a sample before/after:

- Before:
https://spinalcordtoolbox.com/stable/user_section/command-line/deepseg/totalspineseg.html
- After:
https://spinalcordtoolbox--5008.org.readthedocs.build/5008/user_section/command-line/deepseg/totalspineseg.html

## Linked issues

Fixes
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4861.
Fixes
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4974.
Fixes
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4888.

---------

Co-authored-by: Kalum Ost <kalum.ost@polymtl.ca>