Use `ball` of size `5` wherever we dilate single-voxel labels
Previously, the value was `4`, which was OK. Then it was changed to `3`, which is noticeably worse.
Experimentally, `5` seems to preserve even more labels.

I don't think we need this change for the rootlets dilation (do we?), so I kept `3` but added a note instead.

I also didn't update an instance of this in `create_label_z`, because that function has 0 usages, and should
be removed in a follow-up PR.