Refactor the broken link checker (#5047)
Here's a rewrite of the broken link checker, as a standalone bash
script, so that:
- we don't have to repeat mostly the same logic twice; and
- it's checkable by shellcheck.

The new `check-broken-links.sh` script accepts a list of files to check
as its arguments, which can come from either:
- a `find` command for the daily run which checks everything; or
- a `git diff` command for the files touched by any given pull request.

In either case, the exclude list which used to be in
`check_url_blacklist.txt` is now included inline in the
`check-broken-links.sh` script. It could be restored as a file by using
`grep --exclude-from` instead of `grep --exclude`, but I think that
would make a future reader less likely to check the documentation of
`grep --exclude` to see how the exclusions are interpreted (it's
somewhat subtle).

I took the liberty of changing `check-url.sh` so that it takes two real
arguments, rather than having to split around a semicolon.

Fixes #5045.