Switch to `neuropoly` fork of `nnunetv2` to support multi-fold "custom trainer" version of `lesion_ms` (#5027)
## Description

This adds the new MS model to SCT. It is a multi-fold model that outputs
both binary and soft segmentation. Hence, this will require a bit of
change in the code:

1. `nnunetv2` fork
- Switching from `nnunetv2` to `nnunetv2-neuropoly==2.6.2`, which adds
support for custom trainers (along with a few other fixes, such as
returning per-fold logits and also supporting older versions of `torch`)
- Updating `totalspineseg` (which allows for `nnunetv2-neuropoly` to be
installed without conflict)
2. `lesion_ms` multi-fold update
- Replacing the single-fold model URL with a dictionary of 5 fold URLs.
- Updating the existing "multi-seed" model installation code to work
with the `nnunetv2` directory structure.
- Changing the call to the `nnunetv2` inference function to return
multi-fold outputs.
- Adding a function to average the predictions of multi-fold `nnunetv2`
models.
- Adding new model-specific arguments (to allow for single-fold
inference and soft lesion output)

## Linked issues

Fixes #4995 
Fixes #4997.

---------

Co-authored-by: Joshua Newton <joshuacwnewton@gmail.com>