Add new `-yml` arguments to include/exclude both subjects and files based on YAML lists (#5030)
## Description

This PR contains two new arguments (`-include-yml` and `-exclude-yml`)
that allows for YAML files to be passed. The YAML file must contain a
single list containing subjects and/or files, _or_ a dict of lists (as
exported by our QC reports). I based the "single list" format off of the
[`exclude.yml`](https://github.com/sct-pipeline/ukbiobank-spinalcord-csa/blob/04b9dd20bf2ad646e734e3e3effae174d4e451ca/exclude.yml)
file in `ukbiobank`.

The list will be treated differently depending on whether subjects or
files are listed:
- Subjects (e.g. `sub-01` or `sub-01/ses-01`): Identical to our existing
`-include-list` and `-exclude-list` arguments.
- Files (e.g. `sub-01_ses-01_T1w.nii.gz`): The script will still be run,
however an environment variable called `EXCLUDE_FILES` will be created
containing the list of files. That way, the user can choose to skip
specific processing steps within their script.

This follows the conclusions we came to in a previous SCT meeting:
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/3971#issuecomment-3140354514

I've tried to provide example `bash` syntax for using `INCLUDE_FILES` or
`EXCLUDE_FILES` to skip processing. I hope it is clear.

## Linked issues

Fixes #3971.

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>