Update `-centerline` argument to allow for consistent `-discfile` projection (used in aSCOR script) (#5075)
## Description

### Bug

This PR addresses a bug in `sct_compute_ascor` when the `-discfile`
argument is used:

- `sct_compute_ascor` computes the aSCOR by calling
`sct_process_segmentation` on both the SC and canal images, then
performing a calculation using both sets of metrics.
- In order for aggregation to be performed correctly, the SC and Canal
metrics must match in terms of slice + VertLevel (ensuring we don't
accidentally combine rows that have the same slices but different vert
levels, or same vert levels but different slices):

https://github.com/spinalcordtoolbox/spinalcordtoolbox/blob/495a9792c96f317ef34b7f79d1e3fb96d69667f6/spinalcordtoolbox/scripts/sct_compute_ascor.py#L77
- However, when the `-discfile` argument is used, the disc labels are
_independently projected_ onto both the SC seg and canal seg. This can
lead to non-matching sets of slices/vertlevels, meaning the SC and canal
dataframes won't be merged correctly, and output metrics will be
missing.
- To avoid this mismatch, it's just a matter of projecting the labels
onto the same centerline (thus causing the canal and SC seg to be
labeled the same way).

### Fix

- `sct_process_segmentation` already has an argument called
`-angle-corr-centerline` which was introduced to solve a nearly
identical problem: We want to run the command multiple times on
different inputs but use the same centerline (for angle correction).
- We can reuse this argument to _also_ enforce the same centerline for
`-discfile` projection
- Then, in `sct_compute_ascor`, we can use this repurposed `-centerline`
argument to do the same projection for both SC and canal.

### Testing

I started the PR by adding a failing test that demonstrates the missing
outputs. Then, I implemented the fix (as described above), and the test
now passes.

### Discussion

- Since `sct_compute_ascor` now uses a consistent `-centerline` by
default, we _also_ now use consistent angle correction in the aSCOR func
by default too. This changes the output values slightly from the
previous values. (I think this is a desirable side effect, but I wanted
to mention it for transparency.)
- When writing the new test, I noticed that the discfile we were using
before (`t2/labels.nii.gz`) is not a discfile at all, but instead a
"vertebral body label" file with only 2 point labels (meant for
registration). So, I created a disc label image on the fly using
`sct_label_utils`. Should I update the other test to use this disc label
image too?

## Linked issues

Fixes #5071.